bufix

+++ packages/cx-studio/cx_studio/ffmpeg/cx_ffmpeg.py
@@ -1,6 +1,7 @@
 from codecs import ignore_errors
+from collections.abc import Iterable
 import datetime
-from .cx_ffmpeg_exceptions import FFmpegOutputParseError
+from .cx_ffmpeg_exceptions import FFmpegOutputParseError, NoFFmpegExecutableError
 from .cx_ffmpeg_infos import FFmpegCodingInfo, FFmpegFormatInfo, FFmpegProcessInfo
 
 from pathlib import Path
@@ -13,11 +14,12 @@ from functools import wraps
 from datetime import datetime
 from dataclasses import replace
 import threading
+from cx_studio.path_expander import CmdFinder
 
 
 class FFmpeg:
-    def __init__(self, ffmpeg_bin: str | None = None) -> None:
-        self._ffmpeg_bin = ffmpeg_bin or "ffmpeg"
+    def __init__(self, ffmpeg_executable: str | None = None) -> None:
+        self._executable = CmdFinder.which(ffmpeg_executable or "ffmpeg")
         self._signal_handlers = defaultdict(list)
         self._signal_lock = threading.RLock()
         self._cancel_event = threading.Event()
@@ -34,43 +36,54 @@ class FFmpeg:
         return self._running_lock.locked()
 
     def get_basic_info(self, source: str | Path) -> dict:
-        result = {}
+        if not self._executable:
+            raise NoFFmpegExecutableError("No ffmpeg executable found.")
 
-        args = f"-hide_banner -i {source}".split(" ")
+        result = {}
+        streams = []
+
+        # args = f"-hide_banner -i {source}".split(" ")
+        args = [self._executable, "-hide_banner", "-i", source]
 
         with self._running_cond:
-            process = subprocess.Popen(
-                args, executable=self._ffmpeg_bin, stderr=subprocess.PIPE
-            )
-            process.wait()
+            with subprocess.Popen(
+                args,
+                stderr=subprocess.PIPE,
+                text=True,
+                bufsize=1,
+                errors="ignore",
+                universal_newlines=True,
+                encoding="utf-8",
+            ) as process:
+                for line in process.stderr or []:
+                    input_match = re.match(r"Input #0, (.+), from '(.+)':", line)
+                    if input_match:
+                        result["format_name"] = input_match.group(1)
+                        result["file_name"] = input_match.group(2)
+                        continue
 
-        output = process.stderr.read().decode("utf-8") if process.stderr else None
-        if output is None:
-            raise FFmpegOutputParseError("No output from ffmpeg", output)
+                    time_match = re.search(
+                        r"Duration: (.+), start: (.+), bitrate: (\d+\.?\d*\s?\w+)/s",
+                        line,
+                    )
+                    if time_match:
+                        result["duration"] = CxTime.from_timestamp(time_match.group(1))
+                        result["start_time"] = CxTime.from_seconds(
+                            float(time_match.group(2))
+                        )
+                        result["bitrate"] = FileSize.from_string(time_match.group(3))
+                        continue
 
-        streams = []
-        for line in output.split("\n"):
-            input_match = re.match(r"Input #0, (.+), from '(.+)':", line)
-            if input_match:
-                result["format_name"] = input_match.group(1)
-                result["file_name"] = input_match.group(2)
-                continue
+                    streams_match = re.search(r"Stream #0:\d+\s+", line)
+                    if streams_match:
+                        streams.append(line.strip())
+                        continue
 
-            time_match = re.match(
-                r"Duration: (.+), start: (.+), bitrate: (\d+(\.\d+)?\s?\w+/s)", line
-            )
-            if time_match:
-                result["duration"] = CxTime.from_timestamp(time_match.group(1))
-                result["start_time"] = CxTime.from_seconds(float(time_match.group(2)))
-                result["bitrate"] = FileSize.from_string(time_match.group(3))
-                continue
-
-            streams_match = re.match(r"Stream #0:\d+\s+", line)
-            if streams_match:
-                streams.append(line)
-                continue
-
-        result["streams"] = streams
+                process.wait()
+            # process
+        # condition
+        if len(streams) > 0:
+            result["streams"] = streams
         return result
 
     def __emit_signal(self, signal: str, *args, **kwargs):
@@ -92,45 +105,60 @@ class FFmpeg:
 
         return deco
 
-    def run(self, arguments: str | list) -> bool:
-        args = arguments.split(" ") if isinstance(arguments, str) else arguments
+    def run(self, arguments: str | list | Iterable) -> bool:
+        if not self._executable:
+            raise NoFFmpegExecutableError("No ffmpeg executable found.")
+
+        args = [
+            str(x)
+            for x in (arguments.split(" ") if isinstance(arguments, str) else arguments)
+        ]
+        args.insert(0, str(self._executable))
 
         with self._running_cond:
-            process = subprocess.Popen(
-                args, executable=self._ffmpeg_bin, stderr=subprocess.PIPE, text=True
-            )
 
-            basic_process_info = FFmpegProcessInfo(
-                bin=self._ffmpeg_bin, args=args, start_time=datetime.now()
-            )
-            self.__emit_signal("started", basic_process_info)
+            with subprocess.Popen(
+                args,
+                stderr=subprocess.PIPE,
+                text=True,
+                bufsize=1,
+                encoding="utf-8",
+                errors="ignore",
+                universal_newlines=True,
+            ) as process:
 
-            for line in process.stderr:  # type:ignore
-                if basic_process_info.media_duration is None:
-                    match = re.match(r"Duration: (\d{2}:\d{2}:\d{2}\.\d{2})", line)
-                    if match:
-                        duration = CxTime.from_timestamp(match.group(1))
-                        basic_process_info = replace(
-                            basic_process_info, media_duration=duration
+                basic_process_info = FFmpegProcessInfo(
+                    bin=str(self._executable), args=args, start_time=datetime.now()
+                )
+                self.__emit_signal("started", basic_process_info)
+
+                for line in process.stderr or []:
+                    if basic_process_info.media_duration is None:
+                        match = re.match(r"Duration: (\d{2}:\d{2}:\d{2}\.\d{2})", line)
+                        if match:
+                            duration = CxTime.from_timestamp(match.group(1))
+                            basic_process_info = replace(
+                                basic_process_info, media_duration=duration
+                            )
+
+                    line = line.strip()
+                    if line.startswith("frame="):
+                        coding_status = FFmpegCodingInfo.parse_status_line(line)
+                        self.__emit_signal(
+                            "status_updated", basic_process_info, coding_status
                         )
 
-                line = line.strip()
-                if line.startswith("frame="):
-                    coding_status = FFmpegCodingInfo.parse_status_line(line)
-                    self.__emit_signal(
-                        "status_updated", basic_process_info, coding_status
-                    )
-
-                if self._cancel_event.is_set():
-                    process.communicate("q")
-                    process.wait(3)
-                    if process.returncode == 0:
-                        process.kill()
-                        self.__emit_signal("canceled", basic_process_info)
-                    break
-
-            process.wait()
-
+                    if self._cancel_event.is_set():
+                        process.communicate("q")
+                        process.wait(3)
+                        if process.returncode == 0:
+                            process.kill()
+                            self.__emit_signal("canceled", basic_process_info)
+                        break
+                # for
+                process.wait()
+            # process
+        # condition
         basic_process_info = replace(basic_process_info, end_time=datetime.now())
         self.__emit_signal("finished", basic_process_info)
         return process.returncode == 0
